Add 2 new file for Refactoring
