+++ Movie App/Model/Episode.swift
@@ -1,0 +1,15 @@
+//
+//  Episode.swift
+//  Movie App
+//
+//  Created by Amit Kumar on 05/01/19.
+//  Copyright © 2019 IndusUnicorn. All rights reserved.
+//
+
+import Foundation
+
+
+struct Episode: Codable {
+    
+}
+
